Add new dev dependancies for trypescript eg nodemon update scripts in package json added a nodemon json file for development
