+++ nodemon.json
@@ -1,0 +1,5 @@
+{
+    "watch": ["src"],
+    "ext": "ts",
+    "exec": "concurrently \"npx tsc --watch\" \"ts-node src/server.ts\""
+}
